Adding Yarn Workspaces
Closes #19

This PR adds Yarn Workspaces and configures it such that commands in packages are executed from top level.
The idea here is that we never have to `cd` into different directories. Yarn Workspaces also allows us
to share the `node_modules`-folder.

In the future, I see us considering a system that syncs versions between packages and keeps _our_ versions
in sync as well. But that's a story for another time.

Concretely, this PR adds:
- yarn workspaces
- ability to run commands from top level through scripts in `./package.json`
- documentation of the above + how to add new packages
- removes unwanted dependencies in the top level `./package.json`
- renames packages with a `@pengekrukka`-prefix.