Fix type credential types in project
Closes https://github.com/aridder/pengekrukka/issues/81.
We had multiple types of `VerifiableCredential`-type definitions floating around.
I decided to create one type, and base it on Zod schemas. This way, we not only
have one type, but we also avoid having to maintain both a schema and a TS type.

In the process, I discovered that the extra `@pengekrukka/vc-shared` introduced
needless complexity to this project. The above approach would require us to
include Zod in `vc-shared`. This is ab it strange, as Zod schemas does not
make sense in a non-webserver context AFAIK. I decided to remove `vc-shard`;
It was not used by any other package than `webapp`. If we have multiple packages
in the future, we should reintroduce it, but I don't think it is worth the
trouble to keep it.

There are some things still left to be desired (namely figuring out how to
differenciate between credential types in `generateVC` in an elegant way),
but this PR solves our main issue with typing: Every part of the system now
adheres to the `VerifiableCredentialType`. The fact that we now use enums
means that the type validation for `type` is strong as well :)