Patch undefiend personalCredential being returned.
I introduced a bug in #83 that caused `undefined` to be returned to the client as a `PersonCredential`.
While this is only a patch and not a proper fix, I think it's okay in a hackathon-context.
The underlying issue is not solved - The person credential should always be present after boot.